'Warm up' functions before comparing their profiling results'

+++ example/source/qiti_example_include.hpp
@@ -27,8 +27,8 @@ namespace FunctionCallData
 {
 int testHeapAllocation() noexcept;
 int testNoHeapAllocation() noexcept;
-int someWork() noexcept;
-int moreWork() noexcept;
+double someWork() noexcept;
+double moreWork() noexcept;
 } // namespace FunctionCallData
 
 //--------------------------------------------------------------------------
